fixing repository + dto

+++ src/main/java/com/mercadolibre/apicompliance/repository/AuditRepository.java
@@ -1,12 +1,11 @@
 package com.mercadolibre.apicompliance.repository;
 
 import com.mercadolibre.apicompliance.model.Audit;
-import com.mercadolibre.apicompliance.model.Process;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface AuditRepository extends JpaRepository<Audit,Integer> {
+public interface AuditRepository extends JpaRepository<Audit,Long> {
     List<Audit> getAgentsByIp(String ip);
 }
