+++ src/main/java/com/mercadolibre/apicompliance/service/IAuditService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface IAuditService {
     AuditCreatedResponseDTO addNewRecord(AuditRequestDTO agentRequestDTO, String ip);
-    AuditResponseDTO getAuditById(Integer id);
+    AuditResponseDTO getAuditById(Long id);
     List<AuditResponseDTO>  getAuditsByIp(String ip);
     List<ProcessResponseDTO> getProcessByAuditId(Integer id,String filter);
     List<UsersResponseDTO> getUsersByAuditId(Integer id,String filter);
